+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryRootPlugin.kt
@@ -18,9 +18,9 @@ package foundry.gradle
 import com.autonomousapps.DependencyAnalysisExtension
 import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
 import com.osacky.doctor.DoctorExtension
-import com.squareup.moshi.adapter
 import foundry.cli.AppleSiliconCompat
-import foundry.gradle.agp.VersionNumber
+import foundry.common.json.JsonTools
+import foundry.common.versioning.VersionNumber
 import foundry.gradle.avoidance.ComputeAffectedProjectsTask
 import foundry.gradle.avoidance.GenerateAndroidTestProjectPathsTask
 import foundry.gradle.avoidance.GenerateDependencyGraphTask
@@ -29,6 +29,10 @@ import foundry.gradle.develocity.NoOpBuildScanAdapter
 import foundry.gradle.develocity.findAdapter
 import foundry.gradle.lint.DetektTasks
 import foundry.gradle.lint.LintTasks
+import foundry.gradle.properties.StartParameterProperties
+import foundry.gradle.properties.createPropertiesProvider
+import foundry.gradle.properties.setDisallowChanges
+import foundry.gradle.properties.sneakyNull
 import foundry.gradle.stats.ModuleStatsTasks
 import foundry.gradle.tasks.AndroidTestApksTask
 import foundry.gradle.tasks.CoreBootstrapTask
@@ -39,14 +43,8 @@ import foundry.gradle.tasks.KtfmtDownloadTask
 import foundry.gradle.tasks.SortDependenciesDownloadTask
 import foundry.gradle.tasks.robolectric.UpdateRobolectricJarsTask
 import foundry.gradle.unittest.UnitTests
-import foundry.gradle.util.StartParameterProperties
 import foundry.gradle.util.Thermals
 import foundry.gradle.util.ThermalsData
-import foundry.gradle.util.createPropertiesProvider
-import foundry.gradle.util.gitExecProvider
-import foundry.gradle.util.gitVersionProvider
-import foundry.gradle.util.setDisallowChanges
-import foundry.gradle.util.sneakyNull
 import java.util.Locale
 import javax.inject.Inject
 import org.gradle.api.Plugin
@@ -163,7 +161,6 @@ internal class FoundryRootPlugin @Inject constructor(private val buildFeatures: 
         val latestCompileSdkWithSources = foundryProperties.latestCompileSdkWithSources(compileSdk)
         AndroidSourcesConfigurer.patchSdkSources(compileSdk, project, latestCompileSdkWithSources)
       }
-      project.configureGit(foundryProperties)
     }
     project.configureFoundryRootBuildscript(
       foundryProperties.versions.jdk.asProvider(project.providers),
@@ -205,9 +202,7 @@ internal class FoundryRootPlugin @Inject constructor(private val buildFeatures: 
               val text = thermalsLogJsonFile.readText()
               if (text.isNotEmpty()) {
                 try {
-                  thermals =
-                    foundry.gradle.util.JsonTools.MOSHI.adapter<Thermals>()
-                      .fromJson(thermalsLogJsonFile.readText())
+                  thermals = JsonTools.fromJson<Thermals>(thermalsLogJsonFile)
                 } catch (e: Exception) {
                   Logging.getLogger("SGP").error("Failed to parse thermals log", e)
                 }
@@ -418,53 +413,7 @@ internal class FoundryRootPlugin @Inject constructor(private val buildFeatures: 
     return stableKeyword || STABLE_REGEX.matches(version)
   }
 
-  private fun Project.configureGit(foundryProperties: FoundryProperties) {
-    // Only run locally
-    if (!isCi) {
-      foundryProperties.gitHooksFile?.let { hooksPath ->
-        // Configure hooks
-        providers.gitExecProvider("git", "config", "core.hooksPath", hooksPath.canonicalPath).get()
-      }
-
-      val revsFile = foundryProperties.gitIgnoreRevsFile ?: return
-      // "git version 2.24.1"
-      val gitVersion = providers.gitVersionProvider().get()
-      val versionNumber = parseGitVersion(gitVersion)
-      @Suppress(
-        "ReplaceCallWithBinaryOperator"
-      ) // Groovy classes don't seem to export equals() correctly
-      when {
-        versionNumber.equals(VersionNumber.UNKNOWN) -> {
-          logger.lifecycle(
-            "Could not infer git env from '$gitVersion'. This can happen if it's the pre-installed " +
-              "git version from Apple, please consider using a custom git installation from Homebrew or otherwise."
-          )
-        }
-        versionNumber < MIN_GIT_VERSION_FOR_IGNORE_REVS -> {
-          logger.lifecycle(
-            "Current git version ($versionNumber) is too low to use " +
-              "blame.ignoreRevsFile (2.23+). Please consider updating!"
-          )
-        }
-        else -> {
-          logger.debug("Configuring blame.ignoreRevsFile")
-          providers
-            .gitExecProvider("git", "config", "blame.ignoreRevsFile", file(revsFile).canonicalPath)
-            .get()
-        }
-      }
-    }
-  }
-
   private companion object {
-    /**
-     * Minimum supported version of git to use blame.ignoreRevsFile.
-     *
-     * See
-     * https://www.moxio.com/blog/43/ignoring-bulk-change-commits-with-git-blame#git-2.23-to-the-rescue.
-     */
-    val MIN_GIT_VERSION_FOR_IGNORE_REVS = VersionNumber.parse("2.23")
-
     private val STABLE_REGEX = "^[0-9,.v-]+(-android)?(-r)?$".toRegex()
   }
 }