+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/kgp/KgpTasks.kt
@@ -26,8 +26,8 @@ import foundry.gradle.configure
 import foundry.gradle.lint.DetektTasks
 import foundry.gradle.not
 import foundry.gradle.onFirst
+import foundry.gradle.properties.setDisallowChanges
 import foundry.gradle.util.configureKotlinCompilationTask
-import foundry.gradle.util.setDisallowChanges
 import java.io.File
 import java.util.concurrent.atomic.AtomicBoolean
 import org.gradle.api.GradleException
@@ -68,7 +68,7 @@ internal object KgpTasks {
     project.pluginManager.onFirst(KGP_PLUGINS) {
       val kotlinExtension = project.kotlinExtension
       kotlinExtension.apply {
-        kotlinDaemonJvmArgs = foundryTools.globalConfig.kotlinDaemonArgs
+        foundryTools.globalConfig.kotlinDaemonArgs?.let { kotlinDaemonJvmArgs = it }
         foundryProperties.versions.jdk.ifPresent { jdkVersion ->
           jvmToolchain {
             languageVersion.set(JavaLanguageVersion.of(jdkVersion))