+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/lint/DetektTasks.kt
@@ -16,20 +16,20 @@
 package foundry.gradle.lint
 
 import foundry.gradle.FoundryProperties
+import foundry.gradle.artifacts.FoundryArtifact
 import foundry.gradle.artifacts.Publisher
 import foundry.gradle.artifacts.Resolver
-import foundry.gradle.artifacts.SgpArtifact
 import foundry.gradle.avoidance.SkippyArtifacts
 import foundry.gradle.configure
 import foundry.gradle.configureEach
 import foundry.gradle.isRootProject
+import foundry.gradle.properties.setDisallowChanges
+import foundry.gradle.properties.sneakyNull
 import foundry.gradle.register
 import foundry.gradle.tasks.DetektDownloadTask
 import foundry.gradle.tasks.SimpleFileProducerTask
 import foundry.gradle.tasks.SimpleFilesConsumerTask
 import foundry.gradle.tasks.publish
-import foundry.gradle.util.setDisallowChanges
-import foundry.gradle.util.sneakyNull
 import io.gitlab.arturbosch.detekt.Detekt
 import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
 import io.gitlab.arturbosch.detekt.DetektPlugin
@@ -52,7 +52,7 @@ internal object DetektTasks {
         outputFile.setDisallowChanges(project.layout.projectDirectory.file("config/bin/detekt"))
       }
 
-      val resolver = Resolver.interProjectResolver(project, SgpArtifact.SKIPPY_DETEKT)
+      val resolver = Resolver.interProjectResolver(project, FoundryArtifact.SKIPPY_DETEKT)
       SimpleFilesConsumerTask.registerOrConfigure(
         project,
         GLOBAL_CI_DETEKT_TASK_NAME,
@@ -98,7 +98,7 @@ internal object DetektTasks {
 
       val publisher =
         if (affectedProjects == null || project.path in affectedProjects) {
-          Publisher.interProjectPublisher(project, SgpArtifact.SKIPPY_DETEKT)
+          Publisher.interProjectPublisher(project, FoundryArtifact.SKIPPY_DETEKT)
         } else {
           val log = "$LOG Skipping ${project.path}:detekt because it is not affected."
           if (foundryProperties.debug) {