+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/tasks/robolectric/UpdateRobolectricJarsTask.kt
@@ -16,10 +16,10 @@
 package foundry.gradle.tasks.robolectric
 
 import foundry.gradle.FoundryProperties
+import foundry.gradle.properties.setDisallowChanges
 import foundry.gradle.register
 import foundry.gradle.robolectricJars
 import foundry.gradle.tasks.BootstrapTask
-import foundry.gradle.util.setDisallowChanges
 import java.io.File
 import org.gradle.api.DefaultTask
 import org.gradle.api.Project
@@ -159,5 +159,6 @@ private val SDKS =
       DefaultSdk(32, "12.1", "8229987", "REL", 9, -1),
       DefaultSdk(33, "13", "9030017", "Tiramisu", 9, -1),
       DefaultSdk(34, "14", "10818077", "REL", 17, -1),
+      DefaultSdk(35, "15", "12543294", "REL", 17, -1),
     )
     .associateBy { it.apiLevel }