+++ platforms/intellij/compose/src/jvmMain/kotlin/foundry/intellij/compose/aibot/ChatWindowUi.kt
@@ -70,7 +70,11 @@ fun ChatWindowUi(state: ChatScreen.State, modifier: Modifier = Modifier) {
         }
       }
     }
+    if (state.isLoading) {
+      LoadingAnimation(modifier = Modifier.padding(15.dp))
+    }
     ConversationField(
+      isLoading = state.isLoading,
       modifier = Modifier,
       onSendMessage = { userMessage -> state.eventSink(ChatScreen.Event.SendMessage(userMessage)) },
     )
@@ -78,7 +82,11 @@ fun ChatWindowUi(state: ChatScreen.State, modifier: Modifier = Modifier) {
 }
 
 @Composable
-private fun ConversationField(modifier: Modifier = Modifier, onSendMessage: (String) -> Unit) {
+private fun ConversationField(
+  isLoading: Boolean,
+  modifier: Modifier = Modifier,
+  onSendMessage: (String) -> Unit,
+) {
   val textState by remember { mutableStateOf(TextFieldState()) }
   val isTextNotEmpty = textState.text.isNotBlank()
   val (hasStartedConversation, setHasStartedConversation) = remember { mutableStateOf(false) }
@@ -131,13 +139,13 @@ private fun ConversationField(modifier: Modifier = Modifier, onSendMessage: (Str
     Column(Modifier.fillMaxHeight(), verticalArrangement = Arrangement.Center) {
       // button will be disabled if there is no text
       IconButton(
-        modifier = Modifier.padding(4.dp).enabled(isTextNotEmpty),
+        modifier = Modifier.padding(4.dp),
         onClick = {
           if (isTextNotEmpty) {
             sendMessage()
           }
         },
-        enabled = isTextNotEmpty,
+        enabled = isTextNotEmpty && !isLoading,
       ) {
         Icon(
           painter = painterResource("drawable/send.svg"),